feat(web,api): per-role patch upload limits (user/moderator/admin tiers)
Raise single-file + daily upload caps and split them by role:
  user      → 1 GB  file / 1 GB  daily
  moderator → 5 GB  file / 100 GB daily
  admin     → 20 GB file / unlimited daily

Backend:
- constants.UploadTier (MaxFileSize + DailyLimit; UnlimitedDailyUpload sentinel)
  replaces the flat MaxLargeFileSize + 2 daily constants; handler resolves the
  tier from the OAuth roles claim (admin > moderator > user) and threads it
  through Init/Complete. Unlimited tiers skip the daily ceiling but still accrue
  usage (so the quota display stays accurate).
- migration 022: widen user.daily_upload_size int4 → bigint. int4 capped at
  ~2.1 GB, so any file > 2 GB or daily total past 2 GB overflowed the
  `daily_upload_size + size` UPDATE — which silently broke even the old 5 GB
  creator limit. bigint (~9.2 EB) is the permanent type for a byte counter; the
  Go field model.User.DailyUploadSize + MeResponse DTO widen to int64 to match.

Frontend:
- useResourceUpload caps the instant local check at the role's MaxFileSize.
- Publish.vue quota panel: per-role daily limit, "无限制" + hidden bar for admin,
  byte sizes via formatFileSize, per-file hint shows the role's cap.

Per-user enforcement is moyu's; the artifact service still enforces a per-SITE
cap on moyu's OAuth client (artifact_max_file_size / artifact_quota_bytes_daily),
which must be raised to >= 20 GB + generous daily bytes for these tiers to apply.

Requires migration 022.

+++ apps/api/internal/auth/dto/dto.go
@@ -34,7 +34,7 @@ type MeResponse struct {
 	Moemoepoint     int      `json:"moemoepoint"`
 	DailyCheckIn    int      `json:"daily_check_in"`
 	DailyImageCount int      `json:"daily_image_count"`
-	DailyUploadSize int      `json:"daily_upload_size"`
+	DailyUploadSize int64    `json:"daily_upload_size"`
 	FollowerCount   int      `json:"follower_count"`
 	FollowingCount  int      `json:"following_count"`
 }
